+++ sync/fromDisk/questions.sql
@@ -1,7 +1,7 @@
 -- BLOCK insert_question
 INSERT INTO questions
     (uuid,  qid, directory, type,                      title,  options,         client_files,
-        course_id,            grading_method,                       single_variant,
+    partial_credit,   course_id,           grading_method,                       single_variant,
     deleted_at,                      template_directory,
     topic_id,
     external_grading_enabled,
@@ -11,7 +11,7 @@ INSERT INTO questions
     external_grading_timeout)
 (SELECT
     $uuid, $qid, $qid,     $type::enum_question_type, $title, $options::JSONB, $client_files::TEXT[],
-    $course_id::integer, $grading_method::enum_grading_method, $single_variant,
+    $partial_credit, $course_id::integer, $grading_method::enum_grading_method, $single_variant,
     NULL::timestamp with time zone, $template_directory,
     COALESCE((SELECT id FROM topics WHERE name = $topic AND course_id = $course_id), NULL),
     $external_grading_enabled,
@@ -28,6 +28,7 @@ SET
     title = EXCLUDED.title,
     options = EXCLUDED.options,
     client_files = EXCLUDED.client_files,
+    partial_credit = EXCLUDED.partial_credit,
     grading_method = EXCLUDED.grading_method,
     single_variant = EXCLUDED.single_variant,
     template_directory = EXCLUDED.template_directory,